requirements.txt
Flask
psycopg[binary]
psycopg_pool
python-dotenv
flask-cors